feat(device): add Apple Silicon MPS support with auto-detection
- New get_device() utility in networks/__init__.py: cuda > mps > cpu
- All 7 network files now use get_device() instead of hardcoded cuda/cpu
- Fix float64 incompatibility in TD3 warmup noise (MPS requires float32)
- 20 new device tests: detection, all networks use correct device,
  forward+backward on device, full learn steps on device

+++ gsp_rl/src/actors/learning_aids.py
@@ -175,8 +175,8 @@ class NetworkAids(Hyperparameters):
     def TD3_choose_action(self, observation, networks, n_actions):
         if self.time_step < self.warmup:
             mu = T.tensor(np.random.normal(scale = self.noise,
-                                           size = (n_actions,))
-                          ).to(networks['actor'].device)
+                                           size = (n_actions,)),
+                          dtype=T.float).to(networks['actor'].device)
         else:
             state = T.tensor(observation, dtype = T.float).to(networks['actor'].device)
             mu = networks['actor'].forward(state).to(networks['actor'].device)
